--- a/Schema.NET.sln
+++ b/Schema.NET.sln
@@ -94,6 +94,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Schema.NET.Pending", "Sourc
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{AF228B8A-7290-4E26-8506-934C290C6AE3}"
 	ProjectSection(SolutionItems) = preProject
+		Source\Common\Constants.cs = Source\Common\Constants.cs
 		Source\Common\ContactType.cs = Source\Common\ContactType.cs
 		Source\Common\ContextJsonConverter.cs = Source\Common\ContextJsonConverter.cs
 		Source\Common\DateTimeHelper.cs = Source\Common\DateTimeHelper.cs