--- a/Source/Common/Constants.cs
+++ b/Source/Common/Constants.cs
@@ -1,0 +1,18 @@
+namespace Schema.NET
+{
+    /// <summary>
+    /// Various constants for use within Schema.NET
+    /// </summary>
+    public static class Constants
+    {
+        /// <summary>
+        /// The HTTPS URL for Schema.org
+        /// </summary>
+        public const string HttpsSchemaOrgUrl = "https://schema.org";
+
+        /// <summary>
+        /// The HTTPS URL for Schema.org
+        /// </summary>
+        public const string HttpSchemaOrgUrl = "http://schema.org";
+    }
+}