--- a/Source/Common/DateTimeToIso8601DateValuesJsonConverter.cs
+++ b/Source/Common/DateTimeToIso8601DateValuesJsonConverter.cs
@@ -30,11 +30,6 @@ namespace Schema.NET
                 throw new ArgumentNullException(nameof(writer));
             }
 
-            if (value is null)
-            {
-                throw new ArgumentNullException(nameof(value));
-            }
-
             if (options is null)
             {
                 throw new ArgumentNullException(nameof(options));