--- a/Source/Common/DateTimeToIso8601DateValuesJsonConverter.cs
+++ b/Source/Common/DateTimeToIso8601DateValuesJsonConverter.cs
@@ -23,7 +23,7 @@ namespace Schema.NET
         {
 #if NET6_0_OR_GREATER
             ArgumentNullException.ThrowIfNull(writer);
-            ArgumentNullException.ThrowIfNull(serializer);
+            ArgumentNullException.ThrowIfNull(options);
 #else
             if (writer is null)
             {