--- a/Source/Common/EnumHelper.cs
+++ b/Source/Common/EnumHelper.cs
@@ -1,6 +1,7 @@
 namespace Schema.NET
 {
     using System;
+    using System.Diagnostics;
     using System.Diagnostics.CodeAnalysis;
 
     /// <summary>
@@ -8,6 +9,11 @@ namespace Schema.NET
     /// </summary>
     internal static class EnumHelper
     {
+        private const string HttpSchemaOrgUrl = "http://schema.org/";
+        private const int HttpSchemaOrgLength = 18; // equivalent to "http://schema.org/".Length
+        private const string HttpsSchemaOrgUrl = "https://schema.org/";
+        private const int HttpsSchemaOrgLength = 19; // equivalent to "https://schema.org/".Length
+
         /// <summary>
         /// Converts the string representation of the name or numeric value of one or more
         /// enumerated constants to an equivalent enumerated object.
@@ -42,6 +48,50 @@ namespace Schema.NET
             return Enum.TryParse(enumType, value, out result);
 #pragma warning restore IDE0022 // Use expression body for methods
 #endif
+        }
+
+        /// <summary>
+        /// Converts the Schema URI representation of the enum type to an equivalent enumerated object.
+        /// </summary>
+        /// <param name="enumType">The enum type to use for parsing.</param>
+        /// <param name="value">The string representation of the name or numeric value of one or more enumerated constants.</param>
+        /// <param name="result">When this method returns true, an object containing an enumeration constant representing the parsed value.</param>
+        /// <returns><see langword="true"/> if the conversion succeeded; <see langword="false"/> otherwise.</returns>
+        public static bool TryParseEnumFromSchemaUri(
+            Type enumType,
+#if NETCOREAPP3_1_OR_GREATER
+            [NotNullWhen(true)]
+#endif
+            string? value,
+            out object? result)
+        {
+            string? enumString;
+            if (value is not null && value.StartsWith(HttpSchemaOrgUrl, StringComparison.OrdinalIgnoreCase))
+            {
+#pragma warning disable IDE0057 // Use range operator. Need to multi-target.
+                enumString = value.Substring(HttpSchemaOrgLength);
+#pragma warning restore IDE0057 // Use range operator. Need to multi-target.
+            }
+            else if (value is not null && value.StartsWith(HttpsSchemaOrgUrl, StringComparison.OrdinalIgnoreCase))
+            {
+#pragma warning disable IDE0057 // Use range operator. Need to multi-target.
+                enumString = value.Substring(HttpsSchemaOrgLength);
+#pragma warning restore IDE0057 // Use range operator. Need to multi-target.
+            }
+            else
+            {
+                enumString = value;
+            }
+
+            if (TryParse(enumType, enumString, out result))
+            {
+                return true;
+            }
+            else
+            {
+                Debug.WriteLine($"Unable to parse enumeration of type {enumType.FullName} with value {enumString}.");
+                return false;
+            }
         }
     }
 }