--- a/Source/Common/JsonLdObject.cs
+++ b/Source/Common/JsonLdObject.cs
@@ -2,14 +2,12 @@ namespace Schema.NET
 {
     using System;
     using System.Collections.Generic;
-    using System.Runtime.Serialization;
-    using Newtonsoft.Json;
+    using System.Text.Json.Serialization;
 
     /// <summary>
     /// The base JSON-LD object.
     /// See https://json-ld.org/spec/latest/json-ld
     /// </summary>
-    [DataContract]
     public class JsonLdObject : IEquatable<JsonLdObject>
     {
         /// <summary>
@@ -24,7 +22,7 @@ namespace Schema.NET
         /// Simply speaking, a context is used to map terms to IRIs. Terms are case sensitive and any valid string that
         /// is not a reserved JSON-LD keyword can be used as a term.
         /// </summary>
-        [DataMember(Name = "@context", Order = 0)]
+        [JsonPropertyName("@context")]
         [JsonConverter(typeof(ContextJsonConverter))]
         public virtual JsonLdContext Context { get; internal set; } = new JsonLdContext();
 
@@ -32,7 +30,7 @@ namespace Schema.NET
         /// Gets the type, used to uniquely identify things that are being described in the document with IRIs or
         /// blank node identifiers.
         /// </summary>
-        [DataMember(Name = "@type", Order = 1)]
+        [JsonPropertyName("@type")]
         public virtual string? Type { get; }
 
         /// <summary>
@@ -43,7 +41,7 @@ namespace Schema.NET
         /// result in a representation of that node.This may allow an application to retrieve further information about
         /// a node. In JSON-LD, a node is identified using the @id keyword:
         /// </summary>
-        [DataMember(Name = "@id", Order = 2)]
+        [JsonPropertyName("@id")]
         public virtual Uri? Id { get; set; }
 
         /// <summary>