--- a/Tests/Schema.NET.Test/Models/ExternalSchemaModelCustomNamespace.cs
+++ b/Tests/Schema.NET.Test/Models/ExternalSchemaModelCustomNamespace.cs
@@ -1,7 +1,7 @@
 namespace SomeCustomNamespace
 {
     using System.Runtime.Serialization;
-    using Newtonsoft.Json;
+    using System.Text.Json.Serialization;
     using Schema.NET;
 
     public class ExternalSchemaModelCustomNamespace : Thing