… .github/workflows/build.yml
… Benchmarks/Schema.NET.Benchmarks/SchemaBenchmarkBase.cs
… Source/Common/SchemaSerializer.cs
… Source/Common/Thing.Partial.cs
… Source/Common/TimeSpanToISO8601DurationValuesJsonConverter.cs
… Source/Common/ValuesJsonConverter.cs
… Source/Schema.NET.Pending/Schema.NET.Pending.csproj
… Source/Schema.NET/Schema.NET.csproj
… Tests/Schema.NET.Test/DateValuesJsonConverterTest.cs
… Tests/Schema.NET.Test/DurationValuesJsonConverterTest.cs
… Tests/Schema.NET.Test/Examples/JobPostingTest.cs
… Tests/Schema.NET.Test/Examples/RecipeTest.cs
… Tests/Schema.NET.Test/Examples/RestaurantTest.cs
… Tests/Schema.NET.Test/Models/ExternalSchemaModelSharedNamespace.cs
… Tests/Schema.NET.Test/ThingTest.cs
… Tests/Schema.NET.Test/ValuesJsonConverterTest.cs
… Tools/Schema.NET.Tool/SchemaSourceGenerator.cs