@Turnerj 2019-11-20
It doesn't look like some of the functionality that JSON.Net has exists in System.Text.Json, specifically the handling of default values. Currently JSON.Net is configured to ignore default values [however you can't do that with System.Text.Json](https://stackoverflow.com/q/58226377/1676444).

Assuming I can get everything else to work, I think this still might be a deal breaker.