@Turnerj 2019-11-20
A few more opinionated changes in the `Thing` class too with the `ToString` methods.

The short of it is:
- [HTML Escaping is enabled by default in `JsonSerializer`](https://stackoverflow.com/a/58504215/1676444) so I've obsoleted the `ToHtmlEscapedString` method. It can be disabled by using a custom encoder however with it being enabled by default, might be a good time to drop this method - up to you!
- The static `JsonSerializerOptions` need to use a static constructor due to the access level of the `Converters` property (its read only).