@Turnerj 2019-11-20
Currently through getting the tests working, there is a big blocker with: dotnet/corefx#38841

I could manually change every `DataContract` and `DataMember` reference to use the new `JsonPropertyName` property but not sure if that is a good idea.