@RehanSaeed 2019-11-20
This is good stuff. I wondered if there would be road blocks like null value handling problems.

With regards to HTML escaping by default. There are times when you might not want to do that (you're not working in a web context), so we need to provide both methods.

Regarding https://github.com/dotnet/corefx/issues/38841, it may be that we need to wait for .NET Core 3.1 before this PR can be accepted.

Can you not set the order with System.Text.Json at all? I wouldn't mind a complete switch to the new attributes if required. We don't support any other serialization like XML, so there is no benefit to using `DataMember`.