@Turnerj 2019-11-21
As far as I've seen in the source code, no it doesn't seem possible to set the order with the new property. I have a feeling they probably won't have that functionality though as according to the JSON spec, an object is an unordered list of name/value pairs.

As far as I can tell, the serializer will put inherited properties at the end so it is at least still predictable, just that all the tests that do a string comparison will need to have their properties moved around.

I've updated the `Thing.ToString()` methods so the HTML escaping and non-escaping paths are working as intended again.

I'm happy to wait for 3.1 to see if `DataMember` change has landed though can move to `JsonPropertyName` if you want. 🙂