@RehanSaeed 2019-11-23
Happy to move to `JsonPropertyName`, it's just a find and replace to get there.

The ordering is a bigger problem as you have to mess with all of the tests which is a bigger job and much harder to review the PR. If there are no plans to implement ordering in 3.1, then perhaps we should go ahead with it, otherwise wait for 3.1. Thoughts?