@Turnerj 2019-11-28
I really don't see them bringing across ordering to `JsonPropertyName` simply because JSON has no dependency on ordering of object keys normally. At a quick count, there is maybe 36 JSON strings in the tests so it is too much to modify.

It does suck needing to change the tests, less because of the work to do it, more because it is far nicer having the code meet the tests than updating both (code and tests) to meet each other.

I'll look at getting this done soon, even if it is just so we can see if there are any other deal breakers with switching to `System.Text.Json`.