@Turnerj 2019-11-28
Note for self: Need to update `Schema.NET.Tool` - it has a number of string references to `DataMember` and `Newtonsoft` that would need updating.