@RehanSaeed 2019-11-29
Instead of comparing exact string representations in our tests, perhaps we should compare JSON object types from System.Text.Json?