@Turnerj 2019-11-29
That could work! So with Newtonsoft that was `JObject` and in `System.Text.Json` that is `JsonDocument` - it would likely make the code less fragile. It also would have a built in comparer so besides writing up the model, it should fit easily with the test assertions.

The only exception to this is still specifically checking how the `ToString` extensions on `Thing` work. They specifically replace the inner `@context` properties from the JSON and this wouldn't match the normal deserialization process.