@Turnerj 2019-11-30
Hitting a hard error at the moment around `System.Text.Json` either not respecting or ignoring the `ValuesJsonConverter`. It does perform the `CanConvert` check however after that, it does some additional check where it figures it is impossible to turn `String` to `Array`.

It doesn't even call the read method on the `ValuesJsonConverter` which makes this extra weird. If it did call it, I would figure I had a bug in that code. With it not calling it, I'm gonna have to dive into `System.Text.Json` to work out why it even does this.

Fun fact with this though, using `System.Text.Json` 4.6.0 (latest stable) has a harder to understand exception, just saying it couldn't cast `String` to `OneOrMany<String>`. Updating to the latest 4.7.0 preview was able to give me this detail.

```
Message: 
    System.Text.Json.JsonException : The JSON value could not be converted to Schema.NET.OneOrMany`1[System.String]. Path: $.name | LineNumber: 0 | BytePositionInLine: 77.
    ---- System.InvalidOperationException : The requested operation requires an element of type 'Array', but the target element has type 'String'.
  Stack Trace: 
    ThrowHelper.ReThrowWithPath(ReadStack& readStack, Utf8JsonReader& reader, Exception ex)
    JsonSerializer.ReadCore(JsonSerializerOptions options, Utf8JsonReader& reader, ReadStack& readStack)
    JsonSerializer.ReadCore(Type returnType, JsonSerializerOptions options, Utf8JsonReader& reader)
    JsonSerializer.Deserialize(String json, Type returnType, JsonSerializerOptions options)
    JsonSerializer.Deserialize[TValue](String json, JsonSerializerOptions options)
    MusicEventTest.Deserializing_MusicEventJsonLd_ReturnsMusicEvent() line 97
    ----- Inner Stack Trace -----
    JsonDocument.CheckExpectedType(JsonTokenType expected, JsonTokenType actual)
    JsonDocument.GetArrayLength(Int32 index)
    JsonElement.GetArrayLength()
    ValuesJsonConverter.Read(Utf8JsonReader& reader, Type typeToConvert, JsonSerializerOptions options) line 50
    JsonPropertyInfoNotNullableContravariant`4.OnRead(ReadStack& state, Utf8JsonReader& reader)
    JsonPropertyInfo.Read(JsonTokenType tokenType, ReadStack& state, Utf8JsonReader& reader)
    JsonSerializer.ReadCore(JsonSerializerOptions options, Utf8JsonReader& reader, ReadStack& readStack)
```