@Turnerj 2019-11-30
Another note to self: There looks to be a more efficient way to cast raw JSON to value types than using `int.TryParse` etc. This would be used in `ValuesJsonConverter` where many of the direct `TryParse` methods are used directly.

See: https://stackoverflow.com/a/57334833/1676444