@Turnerj 2019-11-30
Had to move to a preview version of `System.Text.Json` as the code was hitting an unavoidable error. It seems like `System.Text.Json` 4.6.0 doesn't understand `OneOrMany` properly and while it uses the `CanConvert` on `ValuesJsonConverter`, it never actually used the converter itself. Anyway, updating to the latest preview avoids that issue.

A few tests are now passing with no change (eg. `MusicEventTest`) though I'm not sure some of these tests are valid. `MusicEventTest` for example deserializes and serializes again a string which wouldn't actually test whether the string JSON is an exact match. It might be worth looking through some of the tests to make sure they are testing the right things. The vast majority seem to be doing that (in at least, most of them are still failing) but some passed where they likely shouldn't have.