@Turnerj 2021-08-14
The commit I've pushed just now should have all tests green! Now it really is fine tuning and having @RehanSaeed review the changes.

Currently with .NET 5, this PR is ~13% faster at serializing while allocating ~14% less. When deserializing, it is ~36% faster while allocating 40% less.

## Main Branch (Book Benchmark)

|      Method |                  Job |              Runtime |     Mean |   Error |  StdDev |      Min |      Max |   Gen 0 | Allocated |
|------------ |--------------------- |--------------------- |---------:|--------:|--------:|---------:|---------:|--------:|----------:|
|   Serialize |             .NET 5.0 |             .NET 5.0 | 131.2 us | 1.98 us | 3.19 us | 126.9 us | 139.0 us | 14.4043 |     45 KB |
| Deserialize |             .NET 5.0 |             .NET 5.0 | 138.7 us | 2.77 us | 3.08 us | 134.7 us | 146.4 us | 25.1465 |     77 KB |
|   Serialize | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 152.5 us | 1.76 us | 1.56 us | 150.4 us | 155.3 us | 14.4043 |     45 KB |
| Deserialize | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 178.8 us | 3.16 us | 2.80 us | 173.3 us | 180.9 us | 26.1230 |     80 KB |

## This PR (Book Benchmark)

|      Method |                  Job |              Runtime |      Mean |    Error |   StdDev |       Min |       Max |   Gen 0 | Allocated |
|------------ |--------------------- |--------------------- |----------:|---------:|---------:|----------:|----------:|--------:|----------:|
|   Serialize |             .NET 5.0 |             .NET 5.0 | 115.14 us | 2.183 us | 2.144 us | 113.10 us | 120.15 us | 12.5732 |     39 KB |
| Deserialize |             .NET 5.0 |             .NET 5.0 |  88.66 us | 1.170 us | 0.977 us |  87.14 us |  89.91 us | 15.0146 |     46 KB |
|   Serialize | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 144.76 us | 2.868 us | 3.522 us | 140.10 us | 151.45 us | 12.4512 |     39 KB |
| Deserialize | .NET Framework 4.7.2 | .NET Framework 4.7.2 | 153.50 us | 1.983 us | 1.855 us | 148.26 us | 156.08 us | 15.8691 |     49 KB |